requirements.txt
openai==1.82.1
google-genai==1.17.0
requests==2.32.3
numpy==2.2.6